🚚 Move `record` arbitrary into /arbitrary
